requirements.txt
streamlit==1.32.2
google-generativeai==0.7.0
python-dotenv==1.0.0
asyncio==3.4.3
mcp>=1.4.0
jira==3.5.2
pandas==2.2.0
psutil>=7.0.0 